LayoutTests/imported/w3c/web-platform-tests/css/motion/offset-supports-calc-expected.txt


FAIL offset-position supports calc assert_equals: expected (string) "calc(20% + 30px) calc(100% - 40px)" but got (undefined) undefined
FAIL offset-path supports calc assert_equals: expected (string) "ray(270deg closest-side)" but got (undefined) undefined
FAIL offset-distance supports calc assert_equals: expected (string) "calc(50% - 100px)" but got (undefined) undefined
FAIL offset-rotate supports calc assert_equals: expected (string) "auto 270deg" but got (undefined) undefined
FAIL offset-anchor supports calc assert_equals: expected (string) "calc(20% + 30px) calc(100% - 40px)" but got (undefined) undefined
